Enhance player management and implement new features
* fix: initialize player limits and expand player creation query with default values and gitignore updates

* fix: waiting login

* refactor: extract database management and account models into a shared ServerData module

* refactor: migrate power and potential fields from JSON to dedicated database columns using long data types

* fix: spawn new mob

* feat: implement SieuHang boss system and TrongTai NPC, and add initial database schema

* refactor: update mission reward system to SMTN, improve maintenance notifications, and add database schema V3

+++ ServerGame/src/main/java/nro/models/clan/Clan.java
@@ -148,7 +148,7 @@ public class Clan {
       return false;
    }
 
-   public void addSMTNClan(Player plOri, double param) {
+   public void addSMTNClan(Player plOri, long param) {
       for (Player pl : this.membersInGame) {
          if (!plOri.equals(pl) && plOri.zone.equals(pl.zone)) {
             Service.getInstance().addSMTN(pl, (byte) 1, param, false);
